requirements.txt
requests
pyyaml
ebooklib
python-docx
beautifulsoup4
PyPDF2
nameparser